created sprites_manager
