+++ worlds_management/Room.py
@@ -1,7 +1,11 @@
+'''
+Manages actual screen display w all sprites and elements, hardcoded values included
+'''
 import pygame
 
-from core import sprites_functions
+import core.sprites_manager
 from resources import image_manager
+print('Room')
 
 pygame.init()
 
@@ -29,16 +33,16 @@ class Room:
         # walls group
         self.walls = pygame.sprite.Group()
         if room_type == "classic":
-            sprites_functions.add_walls(self.walls, room_size)
+            core.sprites_manager.add_walls(self.walls, room_size)
 
         # enemies group
         self.enemies = pygame.sprite.Group()
-        sprites_functions.add_enemies(self.enemies, room_enemies)
+        core.sprites_manager.add_enemies(self.enemies, room_enemies)
 
         # character group
         self.character = pygame.sprite.Group()
         self.character_start_point = [300, 300]
-        sprites_functions.add_character(self.character, self.character_start_point)
+        core.sprites_manager.add_character(self.character, self.character_start_point)
 
         # terrain
         self.terrain_image_start_point = [50, 50]
@@ -65,6 +69,7 @@ class Room:
         self.level_display()
         self.enemy_display()
         self.character_display()
+
 
     def terrain_display(self):
         terrain_image = image_manager.get_terrain_image()
@@ -98,10 +103,10 @@ class Room:
     def money_display(self):
         coin_image = image_manager.get_coin_image()
 
-        from core.main import money
-        money_number = money_font.render(str(money), True, YELLOW)
-        screen.blit(coin_image, self.money_start_point)
-        screen.blit(money_number, [self.money_start_point[0] + 25, self.money_start_point[1]])
+        # from core.main import money
+        #money_number = money_font.render(str(money), True, YELLOW)
+        #screen.blit(coin_image, self.money_start_point)
+        #screen.blit(money_number, [self.money_start_point[0] + 25, self.money_start_point[1]])
 
     def level_display(self):
         for main_character in self.character.sprites():
